backend/pyproject.toml
[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"

[[tool.mypy.overrides]]
module = "awesome_exception.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn.logging.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "amtool"
version = "0.1.0"
description = "church service scheduling"
authors = ["Schwannden Kuo <schwannden@gmail.com>"]
homepage = "https://jethro.schwannden.com"
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[pytest]
mock_use_standalone_module = true
asyncio_mode = "auto"

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
fastapi = "^0.78.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.1.1"}
loguru = "^0.6.0"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
pytz = "^2022.1"
sentry-sdk = "^1.6.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
requests = "^2.28.1"
awesome-exception = "^1.1.0"
google-cloud-firestore = "^2.6.1"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pytest = "^7.1.2"
types-pytz = "^2022.2.1"
bump2version = "^1.0.1"
types-requests = "^2.28.9"
types-PyYAML = "^6.0.12"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"

